# simd4f: Make reciprocal operations 0-safe