## Description
From the main commit messages: (performance measurements omitted)
```
    simd4f: Make reciprocal operations 0-safe
    
    Introduce a well-defined behavior for the reciprocal calculation of
    positive and negative 0. These would now result in positive and
    negative infinity, respectively.
    
    This behavior follows the definition of the IEEE 754 standard.
    
    In the SSE implementation, _mm_rcp_ps() already returned positive or
    negative infinity when used on a positive or negative 0. However, since
    it is then multiplied by 0, we end up with NaN. NaN is then used in the
    next calculations, which finally result in it being returned.
    
    To avoid ending up with NaN, we use a mask created by _mm_cmpeq_ps()
    and a constant, and along with _mm_andnot_ps() we replace infinity with
    0 in the multiplication. After the following subtraction, we end up
    with 2 multiplied by the positive or negative infinity, which finally
    returns the expected result.
    
    The GCC vectorization and scalar implementations are adjusted to return
    infinity when calculating the reciprocal of 0, with copysign() used to
    set the expected sign of the result.
    
    The ARM NEON implementation doesn't require adjusting as it is already
    0-safe.
    
    This comes at a cost of ~11.9% for the SSE implementation, while the
    performance of the GCC vectorization and scalar implementations
    largely remains the same.
```
```
    ray: Remove nudging workaround
    
    Since the reciprocals of positive and negative 0 are now well-defined,
    this workaround should no longer be needed.
```
## Performance
### Tests
`ray-speed`: 50,000,000 ray intersection tests with a box, each using a unique ray.
`simd-speed`: 100,000,000 vector reciprocal calculations, each using a unique vector.
### Results
The following are measurements made for the entire PR. For measurements made for each individual commit, see the commit message.
```
SSE
  BEFORE
    1/2 graphene / ray-speed  OK             1.27s
    2/2 graphene / simd-speed OK             0.42s
  AFTER
    1/2 graphene / ray-speed  OK             1.12s
    2/2 graphene / simd-speed OK             0.47s

GCC VECTORIZATION
  BEFORE
    1/2 graphene / ray-speed  OK             1.32s
    2/2 graphene / simd-speed OK             1.12s
  AFTER
    1/2 graphene / ray-speed  OK             1.27s
    2/2 graphene / simd-speed OK             1.12s

SCALAR
  BEFORE
    1/2 graphene / ray-speed  OK             2.02s
    2/2 graphene / simd-speed OK             1.62s
  AFTER
    1/2 graphene / ray-speed  OK             1.87s
    2/2 graphene / simd-speed OK             1.32s
```
## Status
Ready. The following items were done:
- [x] Add unit tests.
- [x] Measure performance and optimize where possible.
- [x] Test the SSE implementation and adjust where necessary.
- [x] Test the GCC vectorization implementation and adjust where necessary.
- [x] Test the scalar implementation and adjust where necessary.
- [x] Test the ARM NEON implementation and adjust where necessary.

## Dependencies
The following PRs are included in this PR until merged:
- #219 for the GCC vectorization adjustments