@doraskayo 2021-04-02
CI seems to have issues setting up the environment due to `apt-get` failing to download `python3-lxml`:
> Err:16 http://security.ubuntu.com/ubuntu focal-updates/main amd64 python3-lxml amd64 4.5.0-1ubuntu0.2
  404  Not Found [IP: 52.154.174.208 80]
E: Failed to fetch http://security.ubuntu.com/ubuntu/pool/main/l/lxml/python3-lxml_4.5.0-1ubuntu0.2_amd64.deb  404  Not Found [IP: 52.154.174.208 80]
Fetched 4249 kB in 0s (12.3 MB/s)
E: Unable to fetch some archives, maybe run apt-get update or try with --fix-missing?
Error: Process completed with exit code 100.