@doraskayo 2021-04-02
Unfortunately I don't own an ARM device, so I can't test the existing NEON implementation and check whether it needs to be adjusted as well. Looking at the pseudo-code in ARM's documentation and our existing NEON implementation of `graphene_simd4f_reciprocal`, I think it may already behave as expected without modification.

@ebassi, or someone else, I'd appreciate it if you could test this on an ARM device and confirm. I've added unit tests for the expected behavior under `tests/simd.c` for ease of testing.