@doraskayo 2021-04-03
I've used a QEMU user emulated environment to test the NEON implementation, and it does seem to already be 0-safe without modification. However, it's better to confirm this on a bare-metal environment without a translation/emulation layer involved.