@doraskayo 2021-04-03
I enabled NEON for AArch64 on Linux in #221, and included it in this PR for testing purposes. I want to avoid a situation where someone would test this branch on AArch64 hardware and confirm that the NEON implementation doesn't require adjustments without actually running the NEON code-path.