@doraskayo 2021-04-15
I ended up installing Linux on an old `armv7l` phone with NEON support. As expected, the NEON implementation is already 0-safe and requires no additional modifications.

So it seems everything works :grinning: