@doraskayo 2021-04-15
I removed the dependency on #221 as it was only added to ease the testing of the ARM NEON implementation. Now that it was fully tested, it's no longer needed.