@doraskayo 2021-04-17
I made a bunch of before/after performance measurements per commit (see commit messages) and for the entire PR (see PR description).

I think this is ready now. @ebassi, I'd appreciate your review and thoughts.