@ebassi 2021-06-06
GitHub was having a fit every time I tried to run the CI, so I had to merge it manually.

Thanks for the changes!